+++ client/src/app/store/index.ts
@@ -1,2 +1,3 @@
 export { store } from './store';
 export type { IRootState, IAppDispatch, IAppThunk } from './types';
+export { useAppDispatch, useAppSelector } from './hooks';
