+++ client/src/common/components/typography/h2/H2.module.scss
@@ -4,10 +4,3 @@
   line-height: 49px;
   text-transform: uppercase;
 }
-.errorTitle {
-  height: 85vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
