--- client/src/common/helpers/getNormalizedParams.ts
@@ -1,4 +1,0 @@
-export function getNormalizedParams(...params: string[]): string {
-  const normalizedParams = params.filter((param) => param !== '');
-  return normalizedParams.join('&');
-}
