+++ client/src/common/helpers/index.ts
@@ -1,3 +1,2 @@
 export { joinClasses } from './joinClasses';
 export { scrollUp } from './scrollUp';
-export { getNormalizedParams } from './getNormalizedParams';
