+++ client/src/common/hooks/index.ts
@@ -1,3 +1,1 @@
 export { useCatchingEventOutsideElement } from './useCatchingEventOutsideElement';
-export { useAppDispatch } from './useAppDispatch';
-export { useAppSelector } from './useAppSelector';
