--- client/src/common/hooks/useAppDispatch.ts
@@ -1,4 +1,0 @@
-import { IAppDispatch } from '@app/store';
-import { useDispatch } from 'react-redux';
-
-export const useAppDispatch: () => IAppDispatch = useDispatch;
