--- client/src/common/hooks/useAppSelector.ts
@@ -1,4 +1,0 @@
-import { IRootState } from '@app/store';
-import { TypedUseSelectorHook, useSelector } from 'react-redux';
-
-export const useAppSelector: TypedUseSelectorHook<IRootState> = useSelector;
