+++ client/src/features/movies/layouts/collectionMovies/containers/index.ts
@@ -1,1 +1,3 @@
 export { CardMovie } from './CardMovie';
+export { CardsMovie } from './CardsMovie';
+export { SidebarMovie } from './SidebarMovie';
