+++ client/src/features/movies/layouts/collectionMovies/helpers/index.ts
@@ -1,4 +1,4 @@
 export { addStateTooltip } from './addStateTooltip';
 export { getSortingParam } from './getSortingParam';
 export { getGenreParam } from './getGenreParam';
-export { addActiveToMenuItem } from './addActiveToMenuItem';
+export { toggleMenuItem } from './toggleMenuItem';
