+++ client/src/features/movies/layouts/dynamicScreen/containers/index.ts
@@ -1,0 +1,2 @@
+export { DetailMovie } from './DetailMovie';
+export { SearchPanelMovie } from './SearchPanelMovie';
